UI updates and health systems
player can manually gain and lose health by pressing G to gain and H to lose. When the player runs out of the health the game over screen appears and if the user presses R when prompted it will take them to the start of the game. Also the first scene is now the start screen that leads into the main game when the user presses enter